requirements.txt
flask==3.0.3
flask_login
python-dotenv==1.0.1
firebase-admin
requests
bs4
gunicorn
